feat: added problem, try, and either utilities
At some point we will want to be able to handle problems and exceptions
with our code.

The Problem class is useful for reporting what went wrong, whether
that's a simple message or an actual exception.

The Try interface is similar to Optional, but allows for wrapping a
try/catch and returning a Problem if an exception occurs.

The Either interface allows for methods to return two things from a
method instead of just one. Normal kotlin convention is that when a
method returns, it either returns a type, a nullable type, or
throws an exception. Now we can return Either<Problem, T> which
could contain the type, or a reportable Problem.