refactor: simplifying the api
Rather than describing data on how it will be used (request/response) we
should rather describe the data itself (contact/contactWithId). Also
while at some point in the future we may want to have a separation
between our json de/serializer implementation and the data, for now we
shouldn't worry about that as it will add undue complexity and can be
changed easily enough when we get to that point.