feat: added http, logging, and hooked up main
This commit gets us almost completely done: it adds a Ktor http server
for the REST HTTP endpoints, it adds logging, and it hooked up Main to
run the server.

The bulk of the logic is handled by ContactsServer, purposely removing
as much logic out of the server so that there is less to test that
requires spinning up the server.